alga_agent/requirements.txt
requests
googlesearch-python
beautifulsoup4
